backend/requirements-backend.txt
flask
flask-cors
pymupdf
python-docx
pillow
pytesseract
torch==2.9.0
torchvision==0.24.0
timm==1.0.20
transformers==4.57.3
ultralytics==8.3.82
